Modified   ManaPHP/Security/Captcha.php

+++ ManaPHP/Security/Captcha.php
@@ -126,7 +126,7 @@ class Captcha extends Component implements CaptchaInterface
 
         imagefilledrectangle($image, 0, 0, $width, $height, $bgColor);
 
-        $fontFile = $this->alias->resolve($this->_fonts[(int)(mt_rand() % count($this->_fonts))]);
+        $fontFile = $this->alias->resolve($this->_fonts[mt_rand() % count($this->_fonts)]);
 
         $referenceFontSize = min($height, $width / $this->_length);
 
@@ -144,7 +144,7 @@ class Captcha extends Component implements CaptchaInterface
             $points = imagettftext($image, $fontSize, $angle, $x, $y, $fgColor, $fontFile, $code[$i]);
 
             for ($k = 0; $k < $this->_noiseCharCount; $k++) {
-                $letter = $this->_charset[(int)(mt_rand() % strlen($this->_charset))];
+                $letter = $this->_charset[mt_rand() % strlen($this->_charset)];
                 $fgColor = imagecolorallocate($image, mt_rand(0, 240), mt_rand(0, 240), mt_rand(0, 240));
                 imagettftext($image,
                     $fontSize * 0.4 * $this->_rand_amplitude(0.1),
@@ -177,7 +177,7 @@ class Captcha extends Component implements CaptchaInterface
         $image = new \Imagick();
         $draw = new \ImagickDraw();
         $image->newImage($width, $height, new \ImagickPixel('rgb(' . $this->_bgRGB . ')'));
-        $draw->setFont($this->alias->resolve($this->_fonts[(int)(mt_rand() % count($this->_fonts))]));
+        $draw->setFont($this->alias->resolve($this->_fonts[mt_rand() % count($this->_fonts)]));
         $draw->setGravity(\Imagick::GRAVITY_NORTHWEST);
 
         $referenceFontSize = min($height, $width / $this->_length);
@@ -197,7 +197,7 @@ class Captcha extends Component implements CaptchaInterface
             $x += $fontSize * mt_rand(600, 800) / 1000;
 
             for ($k = 0; $k < $this->_noiseCharCount; $k++) {
-                $letter = $this->_charset[(int)(mt_rand() % strlen($this->_charset))];
+                $letter = $this->_charset[mt_rand() % strlen($this->_charset)];
                 $fgPixel->setColor('rgb(' . mt_rand(0, 240) . ',' . mt_rand(0, 240) . ',' . mt_rand(0, 240) . ')');
                 $draw->setFillColor($fgPixel);
                 $draw->setFontSize($fontSize * 0.4 * $this->_rand_amplitude(0.1));
